Solved 70. Climbing Stairs
Initialization: We start by checking a base case. If the input n is equal to 1, we can reach the top in only 1 way, so we return 1.

    Dynamic Programming Array: We create an array called dp of size n + 1 to store the number of ways to reach each step of the staircase.

    Base Cases: Initialize the count for the first two steps (dp[1] and dp[2]). For n = 1, there's 1 way, and for n = 2, there are 2 ways to reach the top (either by taking 1 step twice or 2 steps at once).

    Iterative Calculation: Using a loop starting from step 3 up to n, we calculate the number of ways to reach each step by adding the counts from the previous two steps (dp[i - 1] and dp[i - 2]). The value at dp[i] represents the total ways to reach the i-th step.

    Return Result: Finally, we return dp[n], which holds the number of distinct ways to reach the top of the staircase considering n steps.

Time Complexity: The time complexity of this solution is O(n) because we perform a single loop from 3 to n, calculating the number of ways for each step.

Space Complexity: The space complexity is also O(n) because we create an array of size n + 1 to store the count of ways for each step.