Solved 1791. Find Center of Star Graph
    Input Assumption: The code assumes that the center of the graph is one of the vertices of the first edge in the array edges.

    Comparison: It checks if the first vertex of the first edge (edges[0][0]) is equal to either the first vertex of the second edge (edges[1][0]) or the second vertex of the second edge (edges[1][1]).

    Return First Vertex: If the condition in step 2 is true, the function returns the first vertex of the first edge (edges[0][0]), indicating that it is the center of the graph.

    Return Second Vertex: If the condition in step 2 is false, the function returns the second vertex of the first edge (edges[0][1]), suggesting that it is the center of the graph.

Time Complexity: The time complexity of this code is O(1) because it performs a constant number of comparisons and returns a result without iterating through a collection.

Space Complexity: The space complexity is O(1) as well, as the code uses a fixed amount of memory regardless of the size of the input. It doesn't use any additional data structures that depend on the size of the input.