Solved 136. Single Number
    Initialize a variable: Start by setting a variable named result to 0. This variable will be used to store the cumulative XOR of all elements in the array.

    Iterate through the array: Use a loop to go through each element in the given array (nums).

    XOR operation: For each element, perform the XOR operation between the current value of result and the current element in the array. Update the value of result with the result of this XOR operation.

    Finish iteration: After completing the loop, the result variable will contain the XOR of all elements in the array.

    Return the result: The final value of result is the single number that appears only once in the array.

Time Complexity: The time complexity of this solution is O(n), where n is the length of the input array (nums). This is because the algorithm iterates through each element in the array once.

Space Complexity: The space complexity is O(1) since the algorithm uses only a constant amount of extra space (the result variable) regardless of the input size. It doesn't require additional data structures that grow with the input size.