Solved 231. Power of Two
    Check if n is greater than 0: Ensure that the input number is a positive integer. If it's not positive, it cannot be a power of two.

    Bitwise AND operation: Perform a bitwise AND operation between n and (n - 1). This operation helps us determine if n is a power of two or not.

    Check if the result is zero: If the result of the bitwise AND operation is zero, it means that n is a power of two. This is because powers of two have only one bit set to 1, and when you subtract 1 from a power of two, it flips all the bits to the right of the set bit. So, the bitwise AND of n and (n - 1) will be zero.

Time Complexity: O(1) - The time complexity of this algorithm is constant because it performs a fixed number of operations regardless of the size of the input.

Space Complexity: O(1) - The space complexity is also constant because the algorithm only uses a fixed amount of memory for the input and a few variables, regardless of the input size.