Solved 258. Add Digits
    Start: Begin with the given integer num.
    Loop Condition: While num is greater than or equal to 10:
        Nested Loop: Inside this loop, repeatedly sum the digits of num until it becomes a single-digit number.
            Digit Summing: Extract each digit of num using the modulus operator (% 10) and add it to the sum.
            Digit Extraction: Divide num by 10 to remove the last digit and proceed to the next digit.
        Update num: Assign the value of sum back to num.
    Return: Once num becomes a single-digit number, return it.

Time Complexity: The time complexity of this algorithm is O(log10(num)), where 'num' is the input integer. This complexity arises due to the repeated division of 'num' by 10 in order to extract its digits.

Space Complexity: The space complexity is O(1) because the algorithm only uses a constant amount of extra space regardless of the size of the input integer.